@@ -18,92 +18,77 @@
     "widget"
   ],
   "scripts": {
-    "build:css": "csso src/autocomplete.css -o dist/accessible-autocomplete.min.css",
-    "build:js": "cross-env NODE_ENV=production webpack --progress --display-modules",
-    "build": "run-s 'build:js' 'build:css'",
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --inline --hot --progress",
+    "build": "cross-env NODE_ENV=production webpack --progress",
+    "dev": "cross-env NODE_ENV=development webpack serve",
     "karma:dev": "cross-env NODE_ENV=test karma start test/karma.config.js",
     "karma": "npm run karma:dev -- --single-run",
+    "prepare": "node -e \"try { require('husky').install() } catch (e) {if (e.code !== 'MODULE_NOT_FOUND') throw e}\"",
     "preversion": "npm test",
     "standard": "standard",
+    "pretest": "npm run build",
     "test": "run-p standard karma wdio",
-    "version": "npm run build && git add -A dist",
-    "wdio:test": "cross-env NODE_ENV=test wdio test/wdio.config.js",
-    "wdio": "npm run build wdio:test && git checkout dist/"
-  },
-  "husky": {
-    "hooks": {
-      "pre-push": "npm run build && node scripts/check-staged.js"
-    }
-  },
-  "dependencies": {
-    "preact": "^8.3.1"
+    "version": "git add -A dist",
+    "wdio": "cross-env NODE_ENV=test wdio test/wdio.config.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.5",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-proposal-decorators": "^7.1.2",
-    "@babel/plugin-transform-member-expression-literals": "^7.0.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.1.0",
-    "@babel/plugin-transform-property-literals": "^7.0.0",
-    "@babel/plugin-transform-react-jsx": "^7.0.0",
-    "@babel/preset-env": "^7.1.5",
-    "@babel/register": "^7.0.0",
-    "@wdio/cli": "^6.3.0",
-    "@wdio/local-runner": "^6.3.0",
-    "@wdio/mocha-framework": "^6.3.0",
-    "@wdio/sauce-service": "^6.3.2",
-    "@wdio/selenium-standalone-service": "^6.1.14",
-    "@wdio/spec-reporter": "^6.3.0",
-    "@wdio/static-server-service": "^6.1.14",
-    "@wdio/sync": "^6.3.0",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.4",
-    "babel-plugin-istanbul": "^5.1.0",
-    "chai": "^4.2.0",
-    "chalk": "^2.4.1",
-    "copy-webpack-plugin": "^4.6.0",
-    "coveralls": "^3.0.2",
-    "cross-env": "^5.2.0",
-    "csso-cli": "^1.1.0",
-    "dotenv": "^6.1.0",
-    "husky": "^1.1.3",
-    "karma": "^3.1.1",
-    "karma-chai": "^0.1.0",
-    "karma-chai-sinon": "^0.1.5",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.2",
-    "karma-mocha": "^1.0.1",
+    "@babel/core": "^7.23.7",
+    "@babel/eslint-parser": "^7.23.3",
+    "@babel/plugin-transform-react-jsx": "^7.23.4",
+    "@babel/preset-env": "^7.23.8",
+    "@babel/register": "^7.23.7",
+    "@wdio/cli": "^8.29.0",
+    "@wdio/globals": "^8.29.0",
+    "@wdio/local-runner": "^8.29.0",
+    "@wdio/mocha-framework": "^8.29.0",
+    "@wdio/sauce-service": "^8.29.0",
+    "@wdio/spec-reporter": "^8.29.0",
+    "@wdio/static-server-service": "^8.29.0",
+    "autoprefixer": "^10.4.17",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-istanbul": "^6.1.1",
+    "chai": "^5.0.0",
+    "chalk": "^5.3.0",
+    "core-js": "^3.35.0",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.9.1",
+    "cssnano": "^6.0.3",
+    "cssnano-preset-default": "^6.0.3",
+    "devtools": "^8.29.0",
+    "dotenv": "^16.4.0",
+    "husky": "^8.0.3",
+    "karma": "^6.4.2",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^4.0.0-rc.2",
-    "mocha": "^5.2.0",
+    "karma-sourcemap-loader": "^0.4.0",
+    "karma-webpack": "^5.0.0",
+    "mini-css-extract-plugin": "^2.7.7",
+    "mocha": "^10.2.0",
     "npm-run-all": "^4.1.5",
-    "puppeteer": "^4.0.0",
-    "replace-bundle-webpack-plugin": "^1.0.0",
-    "sinon": "^6.3.5",
-    "sinon-chai": "^3.2.0",
-    "source-map-loader": "^0.2.4",
-    "standard": "^12.0.1",
-    "uglifyjs-webpack-plugin": "^2.0.1",
-    "webdriverio": "^6.3.0",
-    "webpack": "^4.25.1",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-middleware": "^3.4.0",
-    "webpack-dev-server": "^3.1.11",
-    "webpack-sources": "^1.3.0"
+    "postcss": "^8.4.33",
+    "postcss-loader": "^8.0.0",
+    "preact": "^8.5.3",
+    "puppeteer": "^21.7.0",
+    "source-map-loader": "^5.0.0",
+    "standard": "^17.1.0",
+    "terser-webpack-plugin": "^5.3.9",
+    "webdriverio": "^8.28.6",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
-  "browserslist": [
-    ">0.1%",
-    "last 2 Chrome versions",
-    "last 2 Firefox versions",
-    "last 2 Edge versions",
-    "last 2 Samsung versions",
-    "Safari >= 9",
-    "ie 8-11",
-    "iOS >= 9"
-  ],
+  "peerDependencies": {
+    "preact": "^8.0.0"
+  },
+  "peerDependenciesMeta": {
+    "preact": {
+      "optional": true
+    }
+  },
   "standard": {
-    "parser": "babel-eslint"
+    "ignore": [
+      "dist"
+    ],
+    "parser": "@babel/eslint-parser"
   }
 }
