@@ -12,7 +12,7 @@ function accessibleAutocomplete (options) {
 }
 
 const createSimpleEngine = (values) => (query, syncResults) => {
-  var matches = values.filter(r => r.toLowerCase().indexOf(query.toLowerCase()) !== -1)
+  const matches = values.filter(r => r.toLowerCase().indexOf(query.toLowerCase()) !== -1)
   syncResults(matches)
 }
 
@@ -21,7 +21,7 @@ accessibleAutocomplete.enhanceSelectElement = (configurationOptions) => {
 
   // Set defaults.
   if (!configurationOptions.source) {
-    let availableOptions = [].filter.call(configurationOptions.selectElement.options, option => (option.value || configurationOptions.preserveNullOptions))
+    const availableOptions = [].filter.call(configurationOptions.selectElement.options, option => (option.value || configurationOptions.preserveNullOptions))
     configurationOptions.source = availableOptions.map(option => option.textContent || option.innerText)
   }
   configurationOptions.onConfirm = configurationOptions.onConfirm || (query => {
@@ -50,7 +50,7 @@ accessibleAutocomplete.enhanceSelectElement = (configurationOptions) => {
 
   accessibleAutocomplete({
     ...configurationOptions,
-    element: element
+    element
   })
 
   configurationOptions.selectElement.style.display = 'none'
