@@ -1,10 +1,12 @@
-/* global after, describe, before, beforeEach, expect, it */
+/* global after, describe, before, beforeEach, it */
+
+import { expect } from 'chai'
 import { createElement, render } from 'preact' /** @jsx createElement */
 import Autocomplete from '../../src/autocomplete'
 import Status from '../../src/status'
 
 function suggest (query, syncResults) {
-  var results = [
+  const results = [
     'France',
     'Germany',
     'United Kingdom'
@@ -73,8 +75,8 @@ describe('Autocomplete', () => {
       it('renders with an aria-expanded attribute', () => {
         render(<Autocomplete required />, scratch)
 
-        let wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
-        let inputElement = wrapperElement.getElementsByTagName('input')[0]
+        const wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
+        const inputElement = wrapperElement.getElementsByTagName('input')[0]
 
         expect(inputElement.getAttribute('aria-expanded')).to.equal('false')
       })
@@ -82,8 +84,8 @@ describe('Autocomplete', () => {
       it('renders with an aria-describedby attribute', () => {
         render(<Autocomplete id='autocomplete-default' />, scratch)
 
-        let wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
-        let inputElement = wrapperElement.getElementsByTagName('input')[0]
+        const wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
+        const inputElement = wrapperElement.getElementsByTagName('input')[0]
 
         expect(inputElement.getAttribute('aria-describedby')).to.equal('autocomplete-default__assistiveHint')
       })
@@ -92,8 +94,8 @@ describe('Autocomplete', () => {
         it('of value "list", when autoselect is not enabled', () => {
           render(<Autocomplete required />, scratch)
 
-          let wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
-          let inputElement = wrapperElement.getElementsByTagName('input')[0]
+          const wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
+          const inputElement = wrapperElement.getElementsByTagName('input')[0]
 
           expect(inputElement.getAttribute('aria-autocomplete')).to.equal('list')
         })
@@ -101,19 +103,38 @@ describe('Autocomplete', () => {
         it('of value "both", when autoselect is enabled', () => {
           render(<Autocomplete required autoselect />, scratch)
 
-          let wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
-          let inputElement = wrapperElement.getElementsByTagName('input')[0]
+          const wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
+          const inputElement = wrapperElement.getElementsByTagName('input')[0]
 
           expect(inputElement.getAttribute('aria-autocomplete')).to.equal('both')
         })
       })
 
+      it('renders with an aria-labelledby attribute on the menu', () => {
+        render(<Autocomplete menuAttributes={{ 'aria-labelledby': 'autocomplete-default' }} id='autocomplete-default' />, scratch)
+
+        const wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
+        const dropdownElement = wrapperElement.getElementsByTagName('ul')[0]
+        const inputElement = wrapperElement.getElementsByTagName('input')[0]
+
+        expect(dropdownElement.getAttribute('aria-labelledby')).to.equal('autocomplete-default')
+        expect(inputElement.getAttribute('id')).to.equal('autocomplete-default')
+      })
+
+      it('renders with extra class on the input', () => {
+        render(<Autocomplete inputClasses='govuk-input' id='autocomplete-default' />, scratch)
+
+        const inputElement = scratch.getElementsByClassName('autocomplete__input')[0]
+
+        expect(inputElement.getAttribute('class')).to.contain(' govuk-input')
+      })
+
       it('renders with the correct roles', () => {
         render(<Autocomplete required />, scratch)
 
-        let wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
-        let inputElement = wrapperElement.getElementsByTagName('input')[0]
-        let dropdownElement = wrapperElement.getElementsByTagName('ul')[0]
+        const wrapperElement = scratch.getElementsByClassName('autocomplete__wrapper')[0]
+        const inputElement = wrapperElement.getElementsByTagName('input')[0]
+        const dropdownElement = wrapperElement.getElementsByTagName('ul')[0]
 
         expect(inputElement.getAttribute('role')).to.equal('combobox', 'input should have combobox role')
         expect(dropdownElement.getAttribute('role')).to.equal('listbox', 'menu should have listbox role')
@@ -181,6 +202,12 @@ describe('Autocomplete', () => {
         autocomplete.setState({ query: 'f', options: ['France'], menuOpen: true })
         autocomplete.handleInputChange({ target: { value: '' } })
         expect(autocomplete.state.menuOpen).to.equal(false)
+      })
+
+      it('searches with the new term when query length changes', () => {
+        autocomplete.setState({ query: 'fr', options: ['France'] })
+        autocomplete.handleInputChange({ target: { value: 'fb' } })
+        expect(autocomplete.state.options.length).to.equal(0)
       })
 
       it('removes the aria-describedby attribute when query is non empty', () => {
@@ -638,9 +665,9 @@ describe('Status', () => {
       render(<Status />, scratch)
       expect(scratch.innerHTML).to.contain('div')
 
-      let wrapperElement = scratch.getElementsByTagName('div')[0]
-      let ariaLiveA = wrapperElement.getElementsByTagName('div')[0]
-      let ariaLiveB = wrapperElement.getElementsByTagName('div')[1]
+      const wrapperElement = scratch.getElementsByTagName('div')[0]
+      const ariaLiveA = wrapperElement.getElementsByTagName('div')[0]
+      const ariaLiveB = wrapperElement.getElementsByTagName('div')[1]
 
       expect(ariaLiveA.getAttribute('role')).to.equal('status', 'first aria live region should be marked as role=status')
       expect(ariaLiveA.getAttribute('aria-atomic')).to.equal('true', 'first aria live region should be marked as atomic')
@@ -653,7 +680,7 @@ describe('Status', () => {
     describe('behaviour', () => {
       describe('silences aria live announcement', () => {
         it('when a valid choice has been made and the input has focus', (done) => {
-          let status = new Status({
+          const status = new Status({
             ...Status.defaultProps,
             validChoiceMade: true,
             isInFocus: true
@@ -668,7 +695,7 @@ describe('Status', () => {
         })
 
         it('when the input no longer has focus', (done) => {
-          let status = new Status({
+          const status = new Status({
             ...Status.defaultProps,
             validChoiceMade: false,
             isInFocus: false
@@ -684,7 +711,7 @@ describe('Status', () => {
       })
       describe('does not silence aria live announcement', () => {
         it('when a valid choice has not been made and the input has focus', (done) => {
-          let status = new Status({
+          const status = new Status({
             ...Status.defaultProps,
             validChoiceMade: false,
             isInFocus: true
