@@ -1,4 +1,6 @@
-/* global before, beforeEach, after, describe, expect, it */
+/* global before, beforeEach, after, describe, it */
+
+import { expect } from 'chai'
 import accessibleAutocomplete from '../../src/wrapper'
 
 const DEFAULT_OPTIONS = {
@@ -13,7 +15,7 @@ const injectSelectToEnhanceIntoDOM = (element, settings) => {
   settings.options = settings.options || DEFAULT_OPTIONS
   settings.id = settings.id !== undefined ? settings.id : 'location-picker-id'
   settings.name = settings.name !== undefined ? settings.name : 'location-picker-name'
-  var $select = document.createElement('select')
+  const $select = document.createElement('select')
   if (settings.id) {
     $select.id = settings.id
   }
