@@ -1,8 +1,9 @@
 require('@babel/register')({
-  cwd: require('path').resolve(__dirname, '../')
+  rootMode: 'upward'
 })
-var puppeteer = require('puppeteer')
-var webpack = require('../webpack.config.babel.js')[0]
+
+const puppeteer = require('puppeteer')
+const webpackConfig = require('../webpack.config.mjs')
 
 // Use Chrome headless
 process.env.CHROME_BIN = puppeteer.executablePath()
@@ -10,16 +11,8 @@ process.env.CHROME_BIN = puppeteer.executablePath()
 module.exports = function (config) {
   config.set({
     basePath: '../',
-    frameworks: ['mocha', 'chai-sinon'],
-    reporters: ['mocha', 'coverage'],
-    coverageReporter: {
-      dir: 'coverage',
-      reporters: [
-        { type: 'text-summary' },
-        { type: 'html', subdir: 'html' },
-        { type: 'lcov', subdir: 'lcov' }
-      ]
-    },
+    frameworks: ['mocha', 'webpack'],
+    reporters: ['mocha'],
 
     browsers: ['ChromeHeadless'],
 
@@ -33,10 +26,21 @@ module.exports = function (config) {
       '**/*.js': ['sourcemap']
     },
 
-    webpack: webpack,
-    webpackMiddleware: {
-      logLevel: 'error',
-      stats: 'errors-only'
+    webpack: {
+      // Use standalone webpack config [0] rather
+      // than Preact [1] or React [2] configs
+      ...webpackConfig.default[0],
+
+      // Use Karma managed test entry points
+      entry: undefined,
+
+      // Use Karma default `os.tmpdir()` output
+      output: undefined,
+
+      // Suppress webpack performance warnings due to
+      // Karma chunked output and inline source maps
+      performance: { hints: false },
+      stats: { preset: 'errors-only' }
     }
   })
 }
