@@ -1,85 +1,123 @@
 require('dotenv').config()
+require('@babel/register')({
+  rootMode: 'upward'
+})
+
+const { join } = require('path')
+const { cwd } = require('process')
 const puppeteer = require('puppeteer')
-const staticServerPort = process.env.PORT || 4567
-const services = [
-  ['static-server', {
-    folders: [
-      { mount: '/', path: './examples' },
-      { mount: '/dist/', path: './dist' }
-    ],
-    port: staticServerPort
-  }]
+
+const {
+  PORT = 4567,
+  SAUCE_ACCESS_KEY,
+  SAUCE_BUILD_NUMBER,
+  SAUCE_ENABLED,
+  SAUCE_USERNAME
+} = process.env
+
+/**
+ * Browsers for local tests
+ *
+ * @type {RemoteCapabilities}
+ */
+const capabilitiesLocal = [
+  {
+    browserName: 'chrome',
+    'goog:chromeOptions': {
+      args: ['--headless=new'],
+      binary: puppeteer.executablePath()
+    }
+  }
 ]
 
-const sauceEnabled = process.env.SAUCE_ENABLED === 'true'
-const sauceUser = process.env.SAUCE_USERNAME
-const sauceKey = process.env.SAUCE_ACCESS_KEY
-const buildNumber = process.env.TRAVIS_BUILD_NUMBER
-const sauceConfig = sauceEnabled
-  ? {
-    user: sauceUser,
-    key: sauceKey,
-    capabilities: [
-      {
-        browserName: 'chrome',
-        platformName: 'Windows 10',
-        'sauce:options': {
-          build: buildNumber
-        }
-      },
-      {
-        browserName: 'firefox',
-        browserVersion: '55',
-        platformName: 'Windows 10',
-        'sauce:options': {
-          build: buildNumber
-        }
-      },
-      {
-        browserName: 'internet explorer',
-        browserVersion: '11.285',
-        platformName: 'Windows 10',
-        'sauce:options': {
-          build: buildNumber
-        }
-      },
-      {
-        browserName: 'internet explorer',
-        browserVersion: '10',
-        platformName: 'Windows 7',
-        'sauce:options': {
-          build: buildNumber
-        }
-      },
-      /* IE9 on Sauce Labs needs to use legacy JSON Wire Protocol */
-      {
-        browserName: 'internet explorer',
-        version: '9',
-        platform: 'Windows 7',
-        build: buildNumber
-      }
-    ],
-    services: services.concat([['sauce', { sauceConnect: true }]])
-  }
-  : {}
-
-exports.config = Object.assign({
-  outputDir: './logs/',
-  specs: ['./test/integration/**/*.js'],
-  capabilities: [
-    // { browserName: 'firefox' },
-    {
-      browserName: 'chrome',
-      'goog:chromeOptions': {
-        args: ['--headless'],
-        binary: puppeteer.executablePath()
-      }
+/**
+ * Browsers for Sauce Labs tests
+ *
+ * @type {RemoteCapabilities}
+ */
+const capabilitiesSauce = [
+  {
+    browserName: 'chrome',
+    browserVersion: 'latest',
+    platformName: 'Windows 10',
+    'sauce:options': {
+      build: SAUCE_BUILD_NUMBER
     }
-  ],
-  baseUrl: 'http://localhost:' + staticServerPort,
-  screenshotPath: './screenshots/',
-  services: services.concat(['selenium-standalone']),
-  reporters: ['spec'],
+  },
+  {
+    browserName: 'firefox',
+    browserVersion: '55',
+    platformName: 'Windows 10',
+    'sauce:options': {
+      build: SAUCE_BUILD_NUMBER
+    }
+  },
+  {
+    browserName: 'internet explorer',
+    browserVersion: 'latest',
+    platformName: 'Windows 10',
+    'sauce:options': {
+      build: SAUCE_BUILD_NUMBER
+    }
+  }
+]
+
+/**
+ * WebdriverIO config
+ *
+ * @type {Testrunner}
+ */
+exports.config = {
+  user: SAUCE_USERNAME,
+  key: SAUCE_ACCESS_KEY,
+
+  // Use DevTools prototype for Puppeteer
+  automationProtocol: SAUCE_ENABLED === 'true'
+    ? 'webdriver'
+    : 'devtools',
+
+  baseUrl: `http://localhost:${PORT}`,
+
+  capabilities: SAUCE_ENABLED === 'true'
+    ? capabilitiesSauce
+    : capabilitiesLocal,
+
   framework: 'mocha',
-  mochaOpts: { timeout: 30 * 1000 }
-}, sauceConfig)
+  outputDir: join(cwd(), 'logs'),
+  reporters: ['spec'],
+
+  services: [
+    /**
+     * Web server options
+     *
+     * @type {[string, StaticServerOptions]}
+     */
+    ['static-server', {
+      folders: [
+        { mount: '/', path: join(cwd(), 'examples') },
+        { mount: '/dist/', path: join(cwd(), 'dist') }
+      ],
+      port: PORT
+    }],
+
+    /**
+     * Browser testing options
+     *
+     * @type {[string, SauceServiceConfig]}
+     */
+    ['sauce', {
+      // Optionally connect to Sauce Labs
+      sauceConnect: SAUCE_ENABLED === 'true'
+    }]
+  ],
+
+  specs: [join(cwd(), 'test/integration/**/*.js')],
+  waitforTimeout: 30 * 10000
+}
+
+/**
+ * @typedef {import('@wdio/types').Options.Testrunner} Testrunner
+ * @typedef {import('@wdio/types').Capabilities.RemoteCapabilities} RemoteCapabilities
+ * @typedef {import('@wdio/static-server-service').StaticServerOptions} StaticServerOptions
+ * @typedef {import('@wdio/sauce-service').SauceServiceConfig} SauceServiceConfig
+ */
